Adds sharing checkbox form, and styles 'current status' (MH-27)
