+++ myhpom/views/__init__.py
@@ -7,7 +7,7 @@ from myhpom import models
 from myhpom.views.accounts import next_steps
 from myhpom.views.auth import logout
 from myhpom.views.choose_network import choose_network
-from myhpom.views.directive import (upload_current_ad, upload_index, upload_requirements,
+from myhpom.views.upload import (upload_current_ad, upload_index, upload_requirements,
     upload_sharing, upload_submit)
 from myhpom.views.signup import signup
